Bundle jquery.uls and use language-data as dependency
Changes
=======

Introduce rollup based bundler to create single js, css files.
This allows easy usage and no need to include every source code files
explicitly. Fixes issue #326

Add language-data as dependency to the library
Instead of using scripts to clone the language-data repo and then create
our javascript files from it, directly use language-data as dev-dependency.
The final bundle will include language-data. So it is not an external
dependency for bundle.

Approach
========

To avoid any breakage in consumers, no change in jquery plugins were
introduced. No method signature change too. All plugin definitions
moved to src/index.js. index.js also import the language-data
and alias that to $.uls.data.

So, jquery.uls.data.js and jquery.uls.data.utils.js files were deleted.
The scripts to generate this files was also deleted.

All other changes in jquery.uls.*.js is just indendation change
because of the removal of IIFE. The resuling bundle will have IIFE.
IIFE block has to be removed to expose the classes for exporting
as modules.

Examples were updated to refer javascript, css and styles from dist
folder.

Future steps
============

Introduction of build step allows

* To modernize the codebase to new versions of javascript
* To drop jquery dependency and optionally keep the jquery plugins
* Remove grunt based CI system
