+++ examples/index-i18n.html
@@ -4,20 +4,13 @@
 		<meta charset="utf-8">
 		<title>Universal Language Selector</title>
 		<meta name="author" content="Santhosh Thottingal">
-		<link href="../css/jquery.uls.css" rel="stylesheet">
-		<link href="../css/jquery.uls.grid.css" rel="stylesheet">
-		<link href="../css/jquery.uls.lcd.css" rel="stylesheet">
+		<link href="../styles/jquery.uls.css" rel="stylesheet">
 		<!-- demo -->
 		<link href="resources/demo.css" rel="stylesheet">
 		<!-- Libs -->
 		<script src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
 		<!-- Source -->
-		<script src="../src/jquery.uls.data.js"></script>
-		<script src="../src/jquery.uls.data.utils.js"></script>
-		<script src="../src/jquery.uls.lcd.js"></script>
-		<script src="../src/jquery.uls.languagefilter.js"></script>
-		<script src="../src/jquery.uls.core.js"></script>
-		<script src="resources/jquery.i18n.min.js"></script>
+		<script src="../jquery.uls.js"></script>
 		<script>
 			$( function () {
 				var i18n = $.i18n();
