--- scripts/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-	"root": true,
-	"extends": [
-		"wikimedia/server"
-	]
-}
