--- src/jquery.uls.data.js
@@ -1,6104 +1,0 @@
-( function () { function r( e, n, t ) { function o( i, f ) { if ( !n[ i ] ) { if ( !e[ i ] ) { var c = typeof require === 'function' && require; if ( !f && c ) { return c( i, !0 ); } if ( u ) { return u( i, !0 ); } var a = new Error( "Cannot find module '" + i + "'" ); throw a.code = 'MODULE_NOT_FOUND', a; } var p = n[ i ] = { exports: {} }; e[ i ][ 0 ].call( p.exports, function ( r ) { var n = e[ i ][ 1 ][ r ]; return o( n || r ); }, p, p.exports, r, e, n, t ); } return n[ i ].exports; } for ( var u = typeof require === 'function' && require, i = 0; i < t.length; i++ ) { o( t[ i ] ); } return o; } return r; }() )( {
-	1: [ function ( require, module, exports ) {
-		module.exports = ( function ( $ ) {
-			$.uls = $.uls || {};
-			$.uls.data =
-{
-    "languages": {
-        "aa": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Qafár af"
-        ],
-        "ab": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "Аҧсшәа"
-        ],
-        "abe": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Wôbanakiôdwawôgan"
-        ],
-        "abs": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bahasa Ambon"
-        ],
-        "ace": [
-            "Latn",
-            [
-                "AS",
-                "PA"
-            ],
-            "Acèh"
-        ],
-        "acf": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "kwéyòl"
-        ],
-        "ady": [
-            "Cyrl",
-            [
-                "EU",
-                "ME"
-            ],
-            "Адыгабзэ"
-        ],
-        "ady-cyrl": [
-            "ady"
-        ],
-        "ady-latn": [
-            "Latn",
-            [
-                "EU",
-                "ME"
-            ],
-            "Adygabze"
-        ],
-        "aeb": [
-            "aeb-arab"
-        ],
-        "aeb-arab": [
-            "Arab",
-            [
-                "AF"
-            ],
-            "تونسي"
-        ],
-        "aeb-latn": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Tûnsî"
-        ],
-        "af": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Afrikaans"
-        ],
-        "ahr": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "अहिराणी"
-        ],
-        "ak": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Akan"
-        ],
-        "akz": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Albaamo innaaɬiilka"
-        ],
-        "aln": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Gegë"
-        ],
-        "alt": [
-            "Cyrl",
-            [
-                "EU",
-                "AS"
-            ],
-            "алтай тил"
-        ],
-        "am": [
-            "Ethi",
-            [
-                "AF"
-            ],
-            "አማርኛ"
-        ],
-        "ami": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Pangcah"
-        ],
-        "an": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "aragonés"
-        ],
-        "ang": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Ænglisc"
-        ],
-        "ann": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Obolo"
-        ],
-        "anp": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "अङ्गिका"
-        ],
-        "ar": [
-            "Arab",
-            [
-                "AF",
-                "ME"
-            ],
-            "العربية"
-        ],
-        "arc": [
-            "Syrc",
-            [
-                "ME"
-            ],
-            "ܐܪܡܝܐ"
-        ],
-        "arn": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "mapudungun"
-        ],
-        "aro": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Araona"
-        ],
-        "arq": [
-            "Arab",
-            [
-                "AF",
-                "ME"
-            ],
-            "جازايرية"
-        ],
-        "ary": [
-            "Arab",
-            [
-                "AF",
-                "ME"
-            ],
-            "الدارجة"
-        ],
-        "ary-arab": [
-            "ary"
-        ],
-        "ary-latn": [
-            "Latn",
-            [
-                "AF",
-                "ME"
-            ],
-            "ed-dārija"
-        ],
-        "arz": [
-            "Arab",
-            [
-                "AF",
-                "ME"
-            ],
-            "مصرى"
-        ],
-        "as": [
-            "Beng",
-            [
-                "AS"
-            ],
-            "অসমীয়া"
-        ],
-        "ase": [
-            "Sgnw",
-            [
-                "AM"
-            ],
-            "American sign language"
-        ],
-        "ast": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "asturianu"
-        ],
-        "atj": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "atikamekw"
-        ],
-        "atv": [
-            "Cyrl",
-            [
-                "EU",
-                "AS"
-            ],
-            "тÿндÿк алтай тил"
-        ],
-        "av": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "авар"
-        ],
-        "avk": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Kotava"
-        ],
-        "awa": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "अवधी"
-        ],
-        "ay": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Aymar aru"
-        ],
-        "az": [
-            "az-latn"
-        ],
-        "az-arab": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "تۆرکجه"
-        ],
-        "az-latn": [
-            "Latn",
-            [
-                "EU",
-                "ME"
-            ],
-            "azərbaycanca"
-        ],
-        "az-cyrl": [
-            "Cyrl",
-            [
-                "EU",
-                "ME"
-            ],
-            "азәрбајҹанҹа"
-        ],
-        "azb": [
-            "az-arab"
-        ],
-        "azj": [
-            "az-latn"
-        ],
-        "ba": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "башҡортса"
-        ],
-        "ban": [
-            "Latn",
-            [
-                "AS",
-                "PA"
-            ],
-            "Bali"
-        ],
-        "ban-bali": [
-            "Bali",
-            [
-                "AS",
-                "PA"
-            ],
-            "ᬩᬮᬶ"
-        ],
-        "bar": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Boarisch"
-        ],
-        "bas": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "ɓasaá"
-        ],
-        "bat-smg": [
-            "sgs"
-        ],
-        "bbc-latn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Batak Toba"
-        ],
-        "bbc-batk": [
-            "Batk",
-            [
-                "AS"
-            ],
-            "ᯅᯖᯂ᯲ ᯖᯬᯅ"
-        ],
-        "bbc": [
-            "bbc-latn"
-        ],
-        "bcc": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "جهلسری بلوچی"
-        ],
-        "bci": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "wawle"
-        ],
-        "bcl": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bikol Central"
-        ],
-        "bdr": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bajau Sama"
-        ],
-        "be-tarask": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "беларуская (тарашкевіца)"
-        ],
-        "be-x-old": [
-            "be-tarask"
-        ],
-        "be": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "беларуская"
-        ],
-        "bew": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bahasa Betawi"
-        ],
-        "bfa": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Bari"
-        ],
-        "bft": [
-            "Arab",
-            [
-                "AS"
-            ],
-            "بلتی"
-        ],
-        "bfq": [
-            "Taml",
-            [
-                "AS"
-            ],
-            "படகா"
-        ],
-        "bg": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "български"
-        ],
-        "bgn": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "روچ کپتین بلوچی"
-        ],
-        "bh": [
-            "bho"
-        ],
-        "bho": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "भोजपुरी"
-        ],
-        "bi": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Bislama"
-        ],
-        "bjn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Banjar"
-        ],
-        "bkm": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Itaŋikom"
-        ],
-        "blc": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "ItNuxalkmc"
-        ],
-        "bm": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "bamanankan"
-        ],
-        "bn": [
-            "Beng",
-            [
-                "AS"
-            ],
-            "বাংলা"
-        ],
-        "bnn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bunun"
-        ],
-        "bo": [
-            "Tibt",
-            [
-                "AS"
-            ],
-            "བོད་ཡིག"
-        ],
-        "bpy": [
-            "Beng",
-            [
-                "AS"
-            ],
-            "বিষ্ণুপ্রিয়া মণিপুরী"
-        ],
-        "bqi": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "بختیاری"
-        ],
-        "br": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "brezhoneg"
-        ],
-        "brh": [
-            "Latn",
-            [
-                "ME",
-                "AS"
-            ],
-            "Bráhuí"
-        ],
-        "brx": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "बर'"
-        ],
-        "bs": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "bosanski"
-        ],
-        "btm": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Mandailing"
-        ],
-        "bto": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Iriga Bicolano"
-        ],
-        "bug": [
-            "Bugi",
-            [
-                "AS"
-            ],
-            "ᨅᨔ ᨕᨘᨁᨗ"
-        ],
-        "bxr": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "буряад"
-        ],
-        "byn": [
-            "Ethi",
-            [
-                "AF"
-            ],
-            "ብሊን"
-        ],
-        "bzj": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Bileez Kriol"
-        ],
-        "ca": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "català"
-        ],
-        "cak": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Kaqchikel"
-        ],
-        "cbk": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Chavacano de Zamboanga"
-        ],
-        "cbk-zam": [
-            "cbk"
-        ],
-        "ccp": [
-            "Cakm",
-            [
-                "AS"
-            ],
-            "𑄌𑄋𑄴𑄟𑄳𑄦"
-        ],
-        "cdo": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Mìng-dĕ̤ng-ngṳ̄"
-        ],
-        "cdo-latn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Mìng-dĕ̤ng-ngṳ̄ Bàng-uâ-cê"
-        ],
-        "cdo-hani": [
-            "Hani",
-            [
-                "AS"
-            ],
-            "閩東語（漢字）"
-        ],
-        "ce": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "нохчийн"
-        ],
-        "ceb": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Cebuano"
-        ],
-        "ch": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Chamoru"
-        ],
-        "chm": [
-            "mhr"
-        ],
-        "chn": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "chinuk wawa"
-        ],
-        "cho": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Choctaw"
-        ],
-        "chr": [
-            "Cher",
-            [
-                "AM"
-            ],
-            "ᏣᎳᎩ"
-        ],
-        "chy": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Tsetsêhestâhese"
-        ],
-        "ciw": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Anishinaabemowin"
-        ],
-        "cjy": [
-            "cjy-hant"
-        ],
-        "cjy-hans": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "晋语（简化字）"
-        ],
-        "cjy-hant": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "晉語"
-        ],
-        "ckb": [
-            "Arab",
-            [
-                "ME"
-            ],
-            "کوردی"
-        ],
-        "ckt": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "ԓыгъоравэтԓьэн"
-        ],
-        "cnh": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Lai holh"
-        ],
-        "cnr": [
-            "cnr-latn"
-        ],
-        "cnr-cyrl": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "црногорски"
-        ],
-        "cnr-latn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "crnogorski"
-        ],
-        "co": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "corsu"
-        ],
-        "cop": [
-            "Copt",
-            [
-                "AF",
-                "ME"
-            ],
-            "ϯⲙⲉⲧⲣⲉⲙⲛ̀ⲭⲏⲙⲓ"
-        ],
-        "cps": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Capiceño"
-        ],
-        "cr": [
-            "Cans",
-            [
-                "AM"
-            ],
-            "ᓀᐦᐃᔭᐍᐏᐣ"
-        ],
-        "cr-cans": [
-            "cr"
-        ],
-        "cr-latn": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Nēhiyawēwin"
-        ],
-        "crh": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "qırımtatarca"
-        ],
-        "crh-cyrl": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "къырымтатарджа"
-        ],
-        "crh-latn": [
-            "crh"
-        ],
-        "cs": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "čeština"
-        ],
-        "csb": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "kaszëbsczi"
-        ],
-        "cu": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "словѣньскъ \/ ⰔⰎⰑⰂⰡⰐⰠⰔⰍⰟ"
-        ],
-        "cv": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "Чӑвашла"
-        ],
-        "cy": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Cymraeg"
-        ],
-        "da": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "dansk"
-        ],
-        "dag": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "dagbanli"
-        ],
-        "dar": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "дарган"
-        ],
-        "de-at": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Österreichisches Deutsch"
-        ],
-        "de-ch": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Schweizer Hochdeutsch"
-        ],
-        "de-formal": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Deutsch (Sie-Form)"
-        ],
-        "de": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Deutsch"
-        ],
-        "din": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Thuɔŋjäŋ"
-        ],
-        "diq": [
-            "Latn",
-            [
-                "EU",
-                "AS"
-            ],
-            "Zazaki"
-        ],
-        "doi": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "डोगरी"
-        ],
-        "dsb": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "dolnoserbski"
-        ],
-        "dtp": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Dusun Bundu-liwan"
-        ],
-        "dty": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "डोटेली"
-        ],
-        "dv": [
-            "Thaa",
-            [
-                "AS"
-            ],
-            "ދިވެހިބަސް"
-        ],
-        "dz": [
-            "Tibt",
-            [
-                "AS"
-            ],
-            "ཇོང་ཁ"
-        ],
-        "ee": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "eʋegbe"
-        ],
-        "egl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Emiliàn"
-        ],
-        "el": [
-            "Grek",
-            [
-                "EU"
-            ],
-            "Ελληνικά"
-        ],
-        "elm": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Eleme"
-        ],
-        "eml": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "emiliàn e rumagnòl"
-        ],
-        "en-ca": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Canadian English"
-        ],
-        "en-gb": [
-            "Latn",
-            [
-                "EU",
-                "AS",
-                "PA"
-            ],
-            "British English"
-        ],
-        "en-simple": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Simple English"
-        ],
-        "en": [
-            "Latn",
-            [
-                "EU",
-                "AM",
-                "AF",
-                "ME",
-                "AS",
-                "PA",
-                "WW"
-            ],
-            "English"
-        ],
-        "eo": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Esperanto"
-        ],
-        "es-419": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "español de América Latina"
-        ],
-        "es-formal": [
-            "Latn",
-            [
-                "EU",
-                "AM",
-                "AF",
-                "WW"
-            ],
-            "español (formal)"
-        ],
-        "es": [
-            "Latn",
-            [
-                "EU",
-                "AM",
-                "AF",
-                "WW",
-                "PA"
-            ],
-            "español"
-        ],
-        "es-ni": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "español nicaragüense"
-        ],
-        "esu": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Yup'ik"
-        ],
-        "et": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "eesti"
-        ],
-        "eu": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "euskara"
-        ],
-        "ext": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "estremeñu"
-        ],
-        "eya": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "I·ya·q"
-        ],
-        "fa": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "فارسی"
-        ],
-        "fan": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Faŋ"
-        ],
-        "fax": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Fala"
-        ],
-        "ff": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Fulfulde"
-        ],
-        "fi": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "suomi"
-        ],
-        "fil": [
-            "tl"
-        ],
-        "fit": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "meänkieli"
-        ],
-        "fiu-vro": [
-            "vro"
-        ],
-        "fj": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Na Vosa Vakaviti"
-        ],
-        "fkv": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "kvääni"
-        ],
-        "fo": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "føroyskt"
-        ],
-        "fon": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "fɔ̀ngbè"
-        ],
-        "fr": [
-            "Latn",
-            [
-                "EU",
-                "AM",
-                "WW"
-            ],
-            "français"
-        ],
-        "frc": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "français cadien"
-        ],
-        "frp": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "arpetan"
-        ],
-        "frr": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Nordfriisk"
-        ],
-        "fuf": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Fuuta Jalon"
-        ],
-        "fur": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "furlan"
-        ],
-        "fy": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Frysk"
-        ],
-        "ga": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Gaeilge"
-        ],
-        "gaa": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Ga"
-        ],
-        "gag": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Gagauz"
-        ],
-        "gah": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Alekano"
-        ],
-        "gan-hans": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "赣语（简体）"
-        ],
-        "gan-hant": [
-            "gan"
-        ],
-        "gan": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "贛語"
-        ],
-        "gbm": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "गढ़वळि"
-        ],
-        "gbz": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Dari-e Mazdeyasnā"
-        ],
-        "gcf": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Guadeloupean Creole French"
-        ],
-        "gcr": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "kriyòl gwiyannen"
-        ],
-        "gd": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Gàidhlig"
-        ],
-        "gez": [
-            "Ethi",
-            [
-                "AF"
-            ],
-            "ግዕዝ"
-        ],
-        "gl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "galego"
-        ],
-        "gld": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "на̄ни"
-        ],
-        "glk": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "گیلکی"
-        ],
-        "gn": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Avañe'ẽ"
-        ],
-        "gom": [
-            "gom-deva"
-        ],
-        "gom-deva": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "गोंयची कोंकणी"
-        ],
-        "gom-latn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Gõychi Konknni"
-        ],
-        "gor": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bahasa Hulontalo"
-        ],
-        "got": [
-            "Goth",
-            [
-                "EU"
-            ],
-            "𐌲𐌿𐍄𐌹𐍃𐌺"
-        ],
-        "grc": [
-            "Grek",
-            [
-                "EU"
-            ],
-            "Ἀρχαία ἑλληνικὴ"
-        ],
-        "gsw": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Alemannisch"
-        ],
-        "gu": [
-            "Gujr",
-            [
-                "AS"
-            ],
-            "ગુજરાતી"
-        ],
-        "guc": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "wayuunaiki"
-        ],
-        "gum": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Namtrik"
-        ],
-        "gur": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Gurenɛ"
-        ],
-        "guw": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "gungbe"
-        ],
-        "gv": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Gaelg"
-        ],
-        "ha": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Hausa"
-        ],
-        "ha-arab": [
-            "Arab",
-            [
-                "AF"
-            ],
-            "هَوُسَ"
-        ],
-        "ha-latn": [
-            "ha"
-        ],
-        "hai": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "X̱aat Kíl"
-        ],
-        "hak": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Hak-kâ-fa"
-        ],
-        "haw": [
-            "Latn",
-            [
-                "AM",
-                "PA"
-            ],
-            "Hawai`i"
-        ],
-        "he": [
-            "Hebr",
-            [
-                "ME"
-            ],
-            "עברית"
-        ],
-        "hak-hans": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "客家语（简体）"
-        ],
-        "hak-hant": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "客家語（繁體）"
-        ],
-        "hi": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "हिन्दी"
-        ],
-        "hif": [
-            "Latn",
-            [
-                "PA",
-                "AS"
-            ],
-            "Fiji Hindi"
-        ],
-        "hif-deva": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "फ़ीजी हिन्दी"
-        ],
-        "hif-latn": [
-            "hif"
-        ],
-        "hil": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Ilonggo"
-        ],
-        "hne": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "छत्तीसगढ़ी"
-        ],
-        "ho": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Hiri Motu"
-        ],
-        "hoc": [
-            "Wara",
-            [
-                "AS"
-            ],
-            "𑢹𑣉𑣉"
-        ],
-        "hr": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "hrvatski"
-        ],
-        "hrx": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Hunsrik"
-        ],
-        "hsb": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "hornjoserbsce"
-        ],
-        "hsn": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "湘语"
-        ],
-        "ht": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Kreyòl ayisyen"
-        ],
-        "hu-formal": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Magyar (magázó)"
-        ],
-        "hu": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "magyar"
-        ],
-        "hy": [
-            "Armn",
-            [
-                "EU",
-                "ME"
-            ],
-            "հայերեն"
-        ],
-        "hyw": [
-            "Armn",
-            [
-                "EU",
-                "ME"
-            ],
-            "Արեւմտահայերէն"
-        ],
-        "hz": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Otsiherero"
-        ],
-        "ia": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "interlingua"
-        ],
-        "id": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bahasa Indonesia"
-        ],
-        "ie": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Interlingue"
-        ],
-        "ig": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Igbo"
-        ],
-        "ii": [
-            "Yiii",
-            [
-                "AS"
-            ],
-            "ꆇꉙ"
-        ],
-        "ik": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Iñupiak"
-        ],
-        "ike-cans": [
-            "Cans",
-            [
-                "AM"
-            ],
-            "ᐃᓄᒃᑎᑐᑦ"
-        ],
-        "ike-latn": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "inuktitut"
-        ],
-        "ilo": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Ilokano"
-        ],
-        "inh": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "ГӀалгӀай"
-        ],
-        "io": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Ido"
-        ],
-        "is": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "íslenska"
-        ],
-        "it": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "italiano"
-        ],
-        "iu": [
-            "ike-cans"
-        ],
-        "izh": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "ižoran keel"
-        ],
-        "ja": [
-            "Jpan",
-            [
-                "AS"
-            ],
-            "日本語"
-        ],
-        "jam": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Patois"
-        ],
-        "jbo": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "lojban"
-        ],
-        "jdt": [
-            "jdt-cyrl"
-        ],
-        "jdt-cyrl": [
-            "Cyrl",
-            [
-                "EU",
-                "AS"
-            ],
-            "жугьури"
-        ],
-        "jje": [
-            "Kore",
-            [
-                "AS"
-            ],
-            "제주말"
-        ],
-        "jut": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "jysk"
-        ],
-        "jv": [
-            "Latn",
-            [
-                "AS",
-                "PA"
-            ],
-            "Jawa"
-        ],
-        "jv-java": [
-            "Java",
-            [
-                "AS",
-                "PA"
-            ],
-            "ꦗꦮ"
-        ],
-        "ka": [
-            "Geor",
-            [
-                "EU"
-            ],
-            "ქართული"
-        ],
-        "kaa": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Qaraqalpaqsha"
-        ],
-        "kab": [
-            "Latn",
-            [
-                "AF",
-                "EU"
-            ],
-            "Taqbaylit"
-        ],
-        "kac": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Jinghpaw"
-        ],
-        "kbd-cyrl": [
-            "kbd"
-        ],
-        "kbd-latn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Qabardjajəbza"
-        ],
-        "kbd": [
-            "Cyrl",
-            [
-                "EU",
-                "ME"
-            ],
-            "Адыгэбзэ"
-        ],
-        "kbp": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kabɩyɛ"
-        ],
-        "kcg": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Tyap"
-        ],
-        "kea": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kabuverdianu"
-        ],
-        "kg": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kongo"
-        ],
-        "kgp": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Kaingáng"
-        ],
-        "khw": [
-            "Arab",
-            [
-                "ME",
-                "AS"
-            ],
-            "کھوار"
-        ],
-        "ki": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Gĩkũyũ"
-        ],
-        "kiu": [
-            "Latn",
-            [
-                "EU",
-                "ME"
-            ],
-            "Kırmancki"
-        ],
-        "kj": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kwanyama"
-        ],
-        "kjh": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "хакас"
-        ],
-        "kjp": [
-            "Mymr",
-            [
-                "AS"
-            ],
-            "ဖၠုံလိက်"
-        ],
-        "kk": [
-            "kk-cyrl"
-        ],
-        "kk-arab": [
-            "Arab",
-            [
-                "EU",
-                "AS"
-            ],
-            "قازاقشا (تٶتە)"
-        ],
-        "kk-cn": [
-            "kk-arab"
-        ],
-        "kk-cyrl": [
-            "Cyrl",
-            [
-                "EU",
-                "AS"
-            ],
-            "қазақша"
-        ],
-        "kk-kz": [
-            "kk-cyrl"
-        ],
-        "kk-latn": [
-            "Latn",
-            [
-                "EU",
-                "AS",
-                "ME"
-            ],
-            "qazaqşa"
-        ],
-        "kk-tr": [
-            "kk-latn"
-        ],
-        "kl": [
-            "Latn",
-            [
-                "AM",
-                "EU"
-            ],
-            "kalaallisut"
-        ],
-        "km": [
-            "Khmr",
-            [
-                "AS"
-            ],
-            "ភាសាខ្មែរ"
-        ],
-        "kn": [
-            "Knda",
-            [
-                "AS"
-            ],
-            "ಕನ್ನಡ"
-        ],
-        "knn": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "महाराष्ट्रीय कोंकणी"
-        ],
-        "ko-kp": [
-            "Kore",
-            [
-                "AS"
-            ],
-            "조선말"
-        ],
-        "ko": [
-            "Kore",
-            [
-                "AS"
-            ],
-            "한국어"
-        ],
-        "koi": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "перем коми"
-        ],
-        "koy": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Denaakkenaageʼ"
-        ],
-        "kr": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kanuri"
-        ],
-        "krc": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "къарачай-малкъар"
-        ],
-        "kri": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Krio"
-        ],
-        "krj": [
-            "Latn",
-            [
-                "ME",
-                "EU"
-            ],
-            "Kinaray-a"
-        ],
-        "krl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Karjala"
-        ],
-        "ks-arab": [
-            "Arab",
-            [
-                "AS"
-            ],
-            "کٲشُر"
-        ],
-        "ks-deva": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "कॉशुर"
-        ],
-        "ks": [
-            "ks-arab"
-        ],
-        "ksf": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Bafia"
-        ],
-        "ksh": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Ripoarisch"
-        ],
-        "ksw": [
-            "Mymr",
-            [
-                "AS"
-            ],
-            "စှီၤ ကညီကျိာ်"
-        ],
-        "ku": [
-            "ku-latn"
-        ],
-        "ku-arab": [
-            "Arab",
-            [
-                "EU",
-                "ME"
-            ],
-            "كوردي"
-        ],
-        "ku-latn": [
-            "Latn",
-            [
-                "EU",
-                "ME"
-            ],
-            "kurdî"
-        ],
-        "kum": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "къумукъ"
-        ],
-        "kv": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "коми"
-        ],
-        "kw": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "kernowek"
-        ],
-        "ky": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "Кыргызча"
-        ],
-        "la": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Latina"
-        ],
-        "lad": [
-            "lad-latn"
-        ],
-        "lad-latn": [
-            "Latn",
-            [
-                "ME",
-                "EU",
-                "AM"
-            ],
-            "Ladino"
-        ],
-        "lad-hebr": [
-            "Hebr",
-            [
-                "ME",
-                "EU",
-                "AM"
-            ],
-            "לאדינו"
-        ],
-        "lag": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kilaangi"
-        ],
-        "lb": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Lëtzebuergesch"
-        ],
-        "lbe": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "лакку"
-        ],
-        "lez": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "лезги"
-        ],
-        "lfn": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Lingua Franca Nova"
-        ],
-        "lg": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Luganda"
-        ],
-        "li": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Limburgs"
-        ],
-        "lij": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Ligure"
-        ],
-        "liv": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Līvõ kēļ"
-        ],
-        "lki": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "لەکی"
-        ],
-        "lkt": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Lakȟótiyapi"
-        ],
-        "lld": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Ladin"
-        ],
-        "lmo": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "lumbaart"
-        ],
-        "ln": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "lingála"
-        ],
-        "lo": [
-            "Laoo",
-            [
-                "AS"
-            ],
-            "ລາວ"
-        ],
-        "loz": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Silozi"
-        ],
-        "lt": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "lietuvių"
-        ],
-        "lrc": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "لۊری شومالی"
-        ],
-        "ltg": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "latgaļu"
-        ],
-        "lud": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "lüüdi"
-        ],
-        "lus": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Mizo ţawng"
-        ],
-        "lut": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "dxʷləšucid"
-        ],
-        "luz": [
-            "Arab",
-            [
-                "ME"
-            ],
-            "لئری دوٙمینی"
-        ],
-        "lv": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "latviešu"
-        ],
-        "lzh": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "文言"
-        ],
-        "lzz": [
-            "Latn",
-            [
-                "EU",
-                "ME"
-            ],
-            "Lazuri"
-        ],
-        "mad": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Madhurâ"
-        ],
-        "mai": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "मैथिली"
-        ],
-        "map-bms": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Basa Banyumasan"
-        ],
-        "mdf": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "мокшень"
-        ],
-        "mfe": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Morisyen"
-        ],
-        "mg": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Malagasy"
-        ],
-        "mh": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Ebon"
-        ],
-        "mhr": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "олык марий"
-        ],
-        "mi": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Māori"
-        ],
-        "mic": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Mi'kmaq"
-        ],
-        "min": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Minangkabau"
-        ],
-        "miq": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Mískitu"
-        ],
-        "mk": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "македонски"
-        ],
-        "ml": [
-            "Mlym",
-            [
-                "AS",
-                "ME"
-            ],
-            "മലയാളം"
-        ],
-        "mn": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "монгол"
-        ],
-        "mn-cyrl": [
-            "mn"
-        ],
-        "mn-mong": [
-            "mvf"
-        ],
-        "mnc": [
-            "Mong",
-            [
-                "AS"
-            ],
-            "ᠮᠠᠨᠵᡠ ᡤᡳᠰᡠᠨ"
-        ],
-        "mni": [
-            "Mtei",
-            [
-                "AS"
-            ],
-            "ꯃꯤꯇꯩ ꯂꯣꯟ"
-        ],
-        "mni-beng": [
-            "Beng",
-            [
-                "AS"
-            ],
-            "মেইতেই লোন্"
-        ],
-        "mnw": [
-            "Mymr",
-            [
-                "AS"
-            ],
-            "ဘာသာ မန်"
-        ],
-        "mo": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "молдовеняскэ"
-        ],
-        "moe": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "innu-aimun"
-        ],
-        "mr": [
-            "Deva",
-            [
-                "AS",
-                "ME"
-            ],
-            "मराठी"
-        ],
-        "mrh": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Mara"
-        ],
-        "mrj": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "кырык мары"
-        ],
-        "mrv": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Magareva"
-        ],
-        "ms": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bahasa Melayu"
-        ],
-        "ms-arab": [
-            "Arab",
-            [
-                "AS"
-            ],
-            "بهاس ملايو"
-        ],
-        "mt": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Malti"
-        ],
-        "mui": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Musi"
-        ],
-        "mus": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Mvskoke"
-        ],
-        "mvf": [
-            "Mong",
-            [
-                "AS"
-            ],
-            "ᠮᠣᠩᠭᠣᠯ"
-        ],
-        "mwl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Mirandés"
-        ],
-        "mwv": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Behase Mentawei"
-        ],
-        "mww": [
-            "mww-latn"
-        ],
-        "mww-latn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Hmoob Dawb"
-        ],
-        "my": [
-            "Mymr",
-            [
-                "AS"
-            ],
-            "မြန်မာဘာသာ"
-        ],
-        "myv": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "эрзянь"
-        ],
-        "mzn": [
-            "Arab",
-            [
-                "ME",
-                "AS"
-            ],
-            "مازِرونی"
-        ],
-        "na": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Dorerin Naoero"
-        ],
-        "nah": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Nāhuatl"
-        ],
-        "nan": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bân-lâm-gú"
-        ],
-        "nan-hani": [
-            "Hani",
-            [
-                "AS"
-            ],
-            "閩南語（漢字）"
-        ],
-        "nap": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Napulitano"
-        ],
-        "nb": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "norsk (bokmål)"
-        ],
-        "nd": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "siNdebele saseNyakatho"
-        ],
-        "nds-nl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Nedersaksisch"
-        ],
-        "nds": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Plattdüütsch"
-        ],
-        "ne": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "नेपाली"
-        ],
-        "new": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "नेपाल भाषा"
-        ],
-        "ng": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Oshiwambo"
-        ],
-        "nia": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Li Niha"
-        ],
-        "niu": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "ko e vagahau Niuē"
-        ],
-        "njo": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Ao"
-        ],
-        "nl-informal": [
-            "Latn",
-            [
-                "EU",
-                "AM"
-            ],
-            "Nederlands (informeel)"
-        ],
-        "nl": [
-            "Latn",
-            [
-                "EU",
-                "AM"
-            ],
-            "Nederlands"
-        ],
-        "nn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "norsk (nynorsk)"
-        ],
-        "no": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "norsk"
-        ],
-        "nod": [
-            "Thai",
-            [
-                "AS"
-            ],
-            "คำเมือง"
-        ],
-        "nog": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "ногайша"
-        ],
-        "nov": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Novial"
-        ],
-        "nqo": [
-            "Nkoo",
-            [
-                "AF"
-            ],
-            "ߒߞߏ"
-        ],
-        "nr": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "isiNdebele seSewula"
-        ],
-        "nrm": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Nouormand"
-        ],
-        "nso": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Sesotho sa Leboa"
-        ],
-        "nus": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Thok Naath"
-        ],
-        "nv": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Diné bizaad"
-        ],
-        "ny": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Chi-Chewa"
-        ],
-        "nys": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Nyungar"
-        ],
-        "oc": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "occitan"
-        ],
-        "ojb": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Ojibwemowin"
-        ],
-        "oka": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "n̓səl̓xcin̓"
-        ],
-        "olo": [
-            "Latn",
-            [
-                "AS",
-                "EU"
-            ],
-            "livvinkarjala"
-        ],
-        "om": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Oromoo"
-        ],
-        "ood": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "ʼOʼodham ha-ñeʼokĭ"
-        ],
-        "or": [
-            "Orya",
-            [
-                "AS"
-            ],
-            "ଓଡ଼ିଆ"
-        ],
-        "os": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "Ирон"
-        ],
-        "osi": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Using"
-        ],
-        "ota": [
-            "Arab",
-            [
-                "AS",
-                "EU"
-            ],
-            "لسان عثمانى"
-        ],
-        "ovd": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "övdalsk"
-        ],
-        "pa": [
-            "pa-guru"
-        ],
-        "pa-guru": [
-            "Guru",
-            [
-                "AS"
-            ],
-            "ਪੰਜਾਬੀ"
-        ],
-        "pag": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Pangasinan"
-        ],
-        "pam": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Kapampangan"
-        ],
-        "pap": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Papiamentu"
-        ],
-        "pap-aw": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Papiamento"
-        ],
-        "pbb": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Nasa Yuwe"
-        ],
-        "pcd": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Picard"
-        ],
-        "pdc": [
-            "Latn",
-            [
-                "EU",
-                "AM"
-            ],
-            "Deitsch"
-        ],
-        "pdt": [
-            "Latn",
-            [
-                "EU",
-                "AM"
-            ],
-            "Plautdietsch"
-        ],
-        "pfl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Pälzisch"
-        ],
-        "pi": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "पालि"
-        ],
-        "pih": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Norfuk \/ Pitkern"
-        ],
-        "pis": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Pijin"
-        ],
-        "pjt": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Pitjantjatjara"
-        ],
-        "pko": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Pökoot"
-        ],
-        "pl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "polski"
-        ],
-        "pms": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Piemontèis"
-        ],
-        "pnb": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "پنجابی"
-        ],
-        "pnt": [
-            "Grek",
-            [
-                "EU"
-            ],
-            "Ποντιακά"
-        ],
-        "pov": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "guinensi"
-        ],
-        "ppl": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Nawat"
-        ],
-        "prg": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Prūsiskan"
-        ],
-        "prs": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "دری"
-        ],
-        "ps": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "پښتو"
-        ],
-        "pt-br": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "português do Brasil"
-        ],
-        "pt": [
-            "Latn",
-            [
-                "EU",
-                "AM",
-                "AS",
-                "PA",
-                "AF",
-                "WW"
-            ],
-            "português"
-        ],
-        "qu": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Runa Simi"
-        ],
-        "quc": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "K'iche'"
-        ],
-        "qug": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Runa shimi"
-        ],
-        "qwh": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "anqash qichwa"
-        ],
-        "rap": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "arero rapa nui"
-        ],
-        "rcf": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kreol Réyoné"
-        ],
-        "rej": [
-            "Latn",
-            [
-                "AS",
-                "PA"
-            ],
-            "Jang"
-        ],
-        "rgn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Rumagnôl"
-        ],
-        "rhg": [
-            "Rohg",
-            [
-                "AS"
-            ],
-            "𐴌𐴟𐴇𐴥𐴝𐴚𐴒𐴙𐴝"
-        ],
-        "rif": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Tarifit"
-        ],
-        "rki": [
-            "Mymr",
-            [
-                "AS"
-            ],
-            "ရခိုင်"
-        ],
-        "rm": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "rumantsch"
-        ],
-        "rmc": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "romaňi čhib"
-        ],
-        "rmf": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "kaalengo tšimb"
-        ],
-        "rmy": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Romani"
-        ],
-        "rn": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kirundi"
-        ],
-        "ro": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "română"
-        ],
-        "roa-rup": [
-            "rup"
-        ],
-        "roa-tara": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "tarandíne"
-        ],
-        "rtm": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Faeag Rotuma"
-        ],
-        "ru": [
-            "Cyrl",
-            [
-                "EU",
-                "AS",
-                "ME"
-            ],
-            "русский"
-        ],
-        "rue": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "русиньскый"
-        ],
-        "rup": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "armãneashti"
-        ],
-        "ruq": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "Влахесте"
-        ],
-        "ruq-cyrl": [
-            "ruq"
-        ],
-        "ruq-grek": [
-            "Grek",
-            [
-                "EU"
-            ],
-            "Megleno-Romanian (Greek script)"
-        ],
-        "ruq-latn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Vlăheşte"
-        ],
-        "rut": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "мыхаӀбишды"
-        ],
-        "rw": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kinyarwanda"
-        ],
-        "rwr": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "मारवाड़ी"
-        ],
-        "ryu": [
-            "Kana",
-            [
-                "AS"
-            ],
-            "ʔucināguci"
-        ],
-        "sa": [
-            "Deva",
-            [
-                "AS"
-            ],
-            "संस्कृतम्"
-        ],
-        "sah": [
-            "Cyrl",
-            [
-                "EU",
-                "AS"
-            ],
-            "саха тыла"
-        ],
-        "sat": [
-            "Olck",
-            [
-                "AS"
-            ],
-            "ᱥᱟᱱᱛᱟᱲᱤ"
-        ],
-        "saz": [
-            "Saur",
-            [
-                "AS"
-            ],
-            "ꢱꣃꢬꢵꢯ꣄ꢡ꣄ꢬꢵ"
-        ],
-        "sc": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "sardu"
-        ],
-        "scn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "sicilianu"
-        ],
-        "sco": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Scots"
-        ],
-        "sd": [
-            "Arab",
-            [
-                "AS"
-            ],
-            "سنڌي"
-        ],
-        "sdc": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Sassaresu"
-        ],
-        "sdh": [
-            "Arab",
-            [
-                "ME"
-            ],
-            "کوردی خوارگ"
-        ],
-        "se": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "davvisámegiella"
-        ],
-        "ses": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Koyraboro Senni"
-        ],
-        "sei": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Cmique Itom"
-        ],
-        "sg": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Sängö"
-        ],
-        "sgs": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "žemaitėška"
-        ],
-        "sh": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "srpskohrvatski"
-        ],
-        "shi-latn": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Taclḥit"
-        ],
-        "shi-tfng": [
-            "Tfng",
-            [
-                "AF"
-            ],
-            "ⵜⴰⵛⵍⵃⵉⵜ"
-        ],
-        "shi": [
-            "shi-latn"
-        ],
-        "shn": [
-            "Mymr",
-            [
-                "AS"
-            ],
-            "လိၵ်ႈတႆး"
-        ],
-        "shy-latn": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "tacawit"
-        ],
-        "si": [
-            "Sinh",
-            [
-                "AS"
-            ],
-            "සිංහල"
-        ],
-        "simple": [
-            "en-simple"
-        ],
-        "sjd": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "кӣллт са̄мь кӣлл"
-        ],
-        "sje": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "bidumsámegiella"
-        ],
-        "sjo": [
-            "Mong",
-            [
-                "AS"
-            ],
-            "ᠰᡞᠪᡝ ᡤᡞᠰᡠᠨ"
-        ],
-        "sju": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "ubmejesámiengiälla"
-        ],
-        "sk": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "slovenčina"
-        ],
-        "sl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "slovenščina"
-        ],
-        "sli": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Schläsch"
-        ],
-        "slr": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Salırça"
-        ],
-        "sly": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Bahasa Selayar"
-        ],
-        "skr-arab": [
-            "Arab",
-            [
-                "AS"
-            ],
-            "سرائیکی"
-        ],
-        "skr": [
-            "skr-arab"
-        ],
-        "syc": [
-            "Syrc",
-            [
-                "ME"
-            ],
-            "ܣܘܪܝܝܐ"
-        ],
-        "syl": [
-            "Sylo",
-            [
-                "AS"
-            ],
-            "ꠍꠤꠟꠐꠤ"
-        ],
-        "syl-beng": [
-            "Beng",
-            [
-                "AS"
-            ],
-            "সিলেটি"
-        ],
-        "syl-sylo": [
-            "syl"
-        ],
-        "sm": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Gagana Samoa"
-        ],
-        "sma": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "åarjelsaemien"
-        ],
-        "smj": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "julevsámegiella"
-        ],
-        "smn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "anarâškielâ"
-        ],
-        "sms": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "nuõrttsääʹmǩiõll"
-        ],
-        "sn": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "chiShona"
-        ],
-        "so": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Soomaaliga"
-        ],
-        "son": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "soŋay"
-        ],
-        "sq": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "shqip"
-        ],
-        "sr": [
-            "sr-cyrl"
-        ],
-        "sr-ec": [
-            "sr-cyrl"
-        ],
-        "sr-cyrl": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "српски"
-        ],
-        "sr-el": [
-            "sr-latn"
-        ],
-        "sr-latn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "srpski"
-        ],
-        "srn": [
-            "Latn",
-            [
-                "AM",
-                "EU"
-            ],
-            "Sranantongo"
-        ],
-        "ss": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "SiSwati"
-        ],
-        "st": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Sesotho"
-        ],
-        "stq": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Seeltersk"
-        ],
-        "sty": [
-            "Cyrl",
-            [
-                "EU",
-                "AS"
-            ],
-            "себертатар"
-        ],
-        "su": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Sunda"
-        ],
-        "sv": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "svenska"
-        ],
-        "sw": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Kiswahili"
-        ],
-        "swb": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Shikomoro"
-        ],
-        "sxu": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Säggssch"
-        ],
-        "szl": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "ślůnski"
-        ],
-        "szy": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Sakizaya"
-        ],
-        "ta": [
-            "Taml",
-            [
-                "AS"
-            ],
-            "தமிழ்"
-        ],
-        "tay": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Tayal"
-        ],
-        "tcy": [
-            "Knda",
-            [
-                "AS"
-            ],
-            "ತುಳು"
-        ],
-        "te": [
-            "Telu",
-            [
-                "AS"
-            ],
-            "తెలుగు"
-        ],
-        "tet": [
-            "Latn",
-            [
-                "AS",
-                "PA"
-            ],
-            "tetun"
-        ],
-        "tg-cyrl": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "тоҷикӣ"
-        ],
-        "tg-latn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "tojikī"
-        ],
-        "tg": [
-            "tg-cyrl"
-        ],
-        "th": [
-            "Thai",
-            [
-                "AS"
-            ],
-            "ไทย"
-        ],
-        "ti": [
-            "Ethi",
-            [
-                "AF"
-            ],
-            "ትግርኛ"
-        ],
-        "tig": [
-            "Ethi",
-            [
-                "AF"
-            ],
-            "ትግረ"
-        ],
-        "tk": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Türkmençe"
-        ],
-        "tkr": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "ЦӀаӀхна миз"
-        ],
-        "tl": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Tagalog"
-        ],
-        "tly": [
-            "Latn",
-            [
-                "EU",
-                "AS",
-                "ME"
-            ],
-            "tolışi"
-        ],
-        "tly-cyrl": [
-            "Cyrl",
-            [
-                "EU",
-                "AS",
-                "ME"
-            ],
-            "толыши"
-        ],
-        "tmr": [
-            "Hebr",
-            [
-                "ME",
-                "EU",
-                "AM"
-            ],
-            "ארמית בבלית"
-        ],
-        "tn": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Setswana"
-        ],
-        "to": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "lea faka-Tonga"
-        ],
-        "tokipona": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Toki Pona"
-        ],
-        "tpi": [
-            "Latn",
-            [
-                "PA",
-                "AS"
-            ],
-            "Tok Pisin"
-        ],
-        "tr": [
-            "Latn",
-            [
-                "EU",
-                "ME"
-            ],
-            "Türkçe"
-        ],
-        "trp": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Kokborok (Tripuri)"
-        ],
-        "tru": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Ṫuroyo"
-        ],
-        "trv": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Sediq Taroko"
-        ],
-        "ts": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Xitsonga"
-        ],
-        "tsd": [
-            "Grek",
-            [
-                "EU"
-            ],
-            "Τσακωνικά"
-        ],
-        "tt": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "татарча"
-        ],
-        "tt-cyrl": [
-            "tt"
-        ],
-        "tt-latn": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "tatarça"
-        ],
-        "ttt": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "Tati"
-        ],
-        "tum": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "chiTumbuka"
-        ],
-        "tw": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Twi"
-        ],
-        "twd": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Tweants"
-        ],
-        "ty": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "reo tahiti"
-        ],
-        "tyv": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "тыва дыл"
-        ],
-        "tzl": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Talossan"
-        ],
-        "tzm": [
-            "Tfng",
-            [
-                "AF"
-            ],
-            "ⵜⴰⵎⴰⵣⵉⵖⵜ"
-        ],
-        "udm": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "удмурт"
-        ],
-        "ug": [
-            "ug-arab"
-        ],
-        "ug-arab": [
-            "Arab",
-            [
-                "AS"
-            ],
-            "ئۇيغۇرچە"
-        ],
-        "ug-latn": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "uyghurche"
-        ],
-        "ug-cyrl": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "уйғурчә"
-        ],
-        "uk": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "українська"
-        ],
-        "umu": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Huluníixsuwaakan"
-        ],
-        "ur": [
-            "Arab",
-            [
-                "AS",
-                "ME"
-            ],
-            "اردو"
-        ],
-        "uz": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "oʻzbekcha"
-        ],
-        "ve": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Tshivenda"
-        ],
-        "vai": [
-            "Vaii",
-            [
-                "AF"
-            ],
-            "ꕙꔤ"
-        ],
-        "vec": [
-            "Latn",
-            [
-                "EU",
-                "AM"
-            ],
-            "vèneto"
-        ],
-        "vep": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "vepsän kel’"
-        ],
-        "vi": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Tiếng Việt"
-        ],
-        "vls": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "West-Vlams"
-        ],
-        "vmf": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Mainfränkisch"
-        ],
-        "vo": [
-            "Latn",
-            [
-                "WW"
-            ],
-            "Volapük"
-        ],
-        "vot": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Vaďďa"
-        ],
-        "vro": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "võro"
-        ],
-        "wa": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "walon"
-        ],
-        "war": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Winaray"
-        ],
-        "wls": [
-            "Latn",
-            [
-                "PA"
-            ],
-            "Faka'uvea"
-        ],
-        "wo": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Wolof"
-        ],
-        "wuu": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "吴语"
-        ],
-        "xal": [
-            "Cyrl",
-            [
-                "EU"
-            ],
-            "хальмг"
-        ],
-        "xh": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "isiXhosa"
-        ],
-        "xmf": [
-            "Geor",
-            [
-                "EU"
-            ],
-            "მარგალური"
-        ],
-        "xsy": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "SaiSiyat"
-        ],
-        "ydd": [
-            "Hebr",
-            [
-                "AS",
-                "EU"
-            ],
-            "Eastern Yiddish"
-        ],
-        "yi": [
-            "Hebr",
-            [
-                "ME",
-                "EU",
-                "AM"
-            ],
-            "ייִדיש"
-        ],
-        "yo": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "Yorùbá"
-        ],
-        "yrk": [
-            "Cyrl",
-            [
-                "AS"
-            ],
-            "Ненэцяʼ вада"
-        ],
-        "yrl": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "ñe'engatú"
-        ],
-        "yua": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Maaya T'aan"
-        ],
-        "yue": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "粵語"
-        ],
-        "za": [
-            "Latn",
-            [
-                "AS"
-            ],
-            "Vahcuengh"
-        ],
-        "zea": [
-            "Latn",
-            [
-                "EU"
-            ],
-            "Zeêuws"
-        ],
-        "zgh": [
-            "Tfng",
-            [
-                "AF"
-            ],
-            "ⵜⴰⵎⴰⵣⵉⵖⵜ ⵜⴰⵏⴰⵡⴰⵢⵜ"
-        ],
-        "zh": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "中文"
-        ],
-        "zh-classical": [
-            "lzh"
-        ],
-        "zh-cn": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "中文（中国大陆）"
-        ],
-        "zh-hans": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "中文（简体）"
-        ],
-        "zh-hant": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "中文（繁體）"
-        ],
-        "zh-hk": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "中文（香港）"
-        ],
-        "zh-min-nan": [
-            "nan"
-        ],
-        "zh-mo": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "中文（澳門）"
-        ],
-        "zh-my": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "中文（马来西亚）"
-        ],
-        "zh-sg": [
-            "Hans",
-            [
-                "AS"
-            ],
-            "中文（新加坡）"
-        ],
-        "zh-tw": [
-            "Hant",
-            [
-                "AS"
-            ],
-            "中文（台灣）"
-        ],
-        "zh-yue": [
-            "yue"
-        ],
-        "zh-cdo": [
-            "cdo"
-        ],
-        "zu": [
-            "Latn",
-            [
-                "AF"
-            ],
-            "isiZulu"
-        ],
-        "zun": [
-            "Latn",
-            [
-                "AM"
-            ],
-            "Shiwi'ma"
-        ]
-    },
-    "scriptgroups": {
-        "Latin": [
-            "Latn",
-            "Goth"
-        ],
-        "Greek": [
-            "Grek"
-        ],
-        "WestCaucasian": [
-            "Armn",
-            "Geor"
-        ],
-        "Arabic": [
-            "Arab"
-        ],
-        "MiddleEastern": [
-            "Hebr",
-            "Syrc"
-        ],
-        "African": [
-            "Copt",
-            "Ethi",
-            "Nkoo",
-            "Tfng",
-            "Vaii"
-        ],
-        "SouthAsian": [
-            "Beng",
-            "Cakm",
-            "Deva",
-            "Gujr",
-            "Guru",
-            "Knda",
-            "Mlym",
-            "Mtei",
-            "Olck",
-            "Orya",
-            "Rohg",
-            "Saur",
-            "Sinh",
-            "Sylo",
-            "Taml",
-            "Telu",
-            "Tibt",
-            "Thaa",
-            "Wara"
-        ],
-        "Cyrillic": [
-            "Cyrl"
-        ],
-        "CJK": [
-            "Hani",
-            "Hans",
-            "Hant",
-            "Kana",
-            "Kore",
-            "Jpan",
-            "Yiii"
-        ],
-        "SouthEastAsian": [
-            "Bali",
-            "Batk",
-            "Bugi",
-            "Java",
-            "Khmr",
-            "Laoo",
-            "Mymr",
-            "Thai"
-        ],
-        "Mongolian": [
-            "Mong"
-        ],
-        "SignWriting": [
-            "Sgnw"
-        ],
-        "NativeAmerican": [
-            "Cher",
-            "Cans"
-        ],
-        "Special": [
-            "Zyyy"
-        ]
-    },
-    "rtlscripts": [
-        "Arab",
-        "Hebr",
-        "Syrc",
-        "Nkoo",
-        "Rohg",
-        "Thaa"
-    ],
-    "regiongroups": {
-        "WW": 1,
-        "SP": 1,
-        "AM": 2,
-        "EU": 3,
-        "ME": 3,
-        "AF": 3,
-        "AS": 4,
-        "PA": 4
-    },
-    "territories": {
-        "AC": [
-            "en"
-        ],
-        "AD": [
-            "ca",
-            "es",
-            "fr"
-        ],
-        "AE": [
-            "ar",
-            "en",
-            "ml",
-            "ps",
-            "fa"
-        ],
-        "AF": [
-            "fa",
-            "ps",
-            "uz",
-            "tk",
-            "bgn",
-            "ug-arab",
-            "kk-arab",
-            "kk-cyrl"
-        ],
-        "AG": [
-            "en",
-            "pt"
-        ],
-        "AI": [
-            "en"
-        ],
-        "AL": [
-            "sq",
-            "el",
-            "mk"
-        ],
-        "AM": [
-            "hy",
-            "ku-latn",
-            "az-latn"
-        ],
-        "AO": [
-            "pt",
-            "ln"
-        ],
-        "AR": [
-            "es",
-            "en",
-            "cy",
-            "gn"
-        ],
-        "AS": [
-            "sm",
-            "en"
-        ],
-        "AT": [
-            "de",
-            "bar",
-            "en",
-            "fr",
-            "it",
-            "hr",
-            "sl",
-            "hu"
-        ],
-        "AU": [
-            "en",
-            "zh-hant",
-            "zh",
-            "it"
-        ],
-        "AW": [
-            "nl",
-            "pap",
-            "en"
-        ],
-        "AX": [
-            "sv"
-        ],
-        "AZ": [
-            "az-latn",
-            "az-cyrl",
-            "tly",
-            "ku-latn",
-            "ttt",
-            "tkr"
-        ],
-        "BA": [
-            "bs",
-            "en",
-            "hr",
-            "sr-cyrl",
-            "sr-latn"
-        ],
-        "BB": [
-            "en"
-        ],
-        "BD": [
-            "bn",
-            "en",
-            "syl",
-            "ccp",
-            "my",
-            "mni"
-        ],
-        "BE": [
-            "en",
-            "nl",
-            "fr",
-            "de",
-            "vls",
-            "wa"
-        ],
-        "BF": [
-            "fr",
-            "ff"
-        ],
-        "BG": [
-            "bg",
-            "en",
-            "ru",
-            "tr",
-            "de"
-        ],
-        "BH": [
-            "ar",
-            "ml"
-        ],
-        "BI": [
-            "rn",
-            "fr",
-            "sw",
-            "en"
-        ],
-        "BJ": [
-            "fr",
-            "fon",
-            "yo"
-        ],
-        "BL": [
-            "fr"
-        ],
-        "BM": [
-            "en"
-        ],
-        "BN": [
-            "ms",
-            "zh-hant",
-            "zh",
-            "ms-arab",
-            "en"
-        ],
-        "BO": [
-            "es",
-            "qu",
-            "ay",
-            "gn",
-            "aro"
-        ],
-        "BQ": [
-            "pap",
-            "nl"
-        ],
-        "BR": [
-            "pt",
-            "en",
-            "de",
-            "it",
-            "ja",
-            "es",
-            "kgp",
-            "ko",
-            "yrl"
-        ],
-        "BS": [
-            "en"
-        ],
-        "BT": [
-            "dz",
-            "ne",
-            "en"
-        ],
-        "BW": [
-            "en",
-            "tn",
-            "af"
-        ],
-        "BY": [
-            "be",
-            "ru"
-        ],
-        "BZ": [
-            "en",
-            "es"
-        ],
-        "CA": [
-            "en",
-            "fr",
-            "it",
-            "de",
-            "pdt",
-            "cr",
-            "yi",
-            "ike-cans",
-            "moe",
-            "atj"
-        ],
-        "CC": [
-            "ms-arab",
-            "ms",
-            "en"
-        ],
-        "CD": [
-            "sw",
-            "fr",
-            "ln",
-            "kg",
-            "rw"
-        ],
-        "CF": [
-            "fr",
-            "sg",
-            "ln"
-        ],
-        "CG": [
-            "fr",
-            "ln"
-        ],
-        "CH": [
-            "de",
-            "gsw",
-            "en",
-            "fr",
-            "it",
-            "lmo",
-            "pt",
-            "rm"
-        ],
-        "CI": [
-            "fr",
-            "bci"
-        ],
-        "CK": [
-            "en"
-        ],
-        "CL": [
-            "es",
-            "en",
-            "arn"
-        ],
-        "CM": [
-            "fr",
-            "en",
-            "ff",
-            "bkm",
-            "bas",
-            "ar",
-            "ksf",
-            "ha-arab",
-            "ha"
-        ],
-        "CN": [
-            "zh",
-            "wuu",
-            "yue",
-            "hsn",
-            "hak",
-            "nan",
-            "gan",
-            "ii",
-            "ug-arab",
-            "za",
-            "mvf",
-            "mn",
-            "bo",
-            "ko",
-            "kk-arab",
-            "kk-cyrl",
-            "ky",
-            "en",
-            "ru",
-            "vi",
-            "uz",
-            "lzh"
-        ],
-        "CO": [
-            "es",
-            "guc"
-        ],
-        "CR": [
-            "es"
-        ],
-        "CU": [
-            "es"
-        ],
-        "CV": [
-            "kea",
-            "pt"
-        ],
-        "CW": [
-            "pap",
-            "nl",
-            "es"
-        ],
-        "CX": [
-            "en"
-        ],
-        "CY": [
-            "el",
-            "en",
-            "tr",
-            "fr",
-            "hy",
-            "ar"
-        ],
-        "CZ": [
-            "cs",
-            "en",
-            "sk",
-            "de",
-            "pl"
-        ],
-        "DE": [
-            "de",
-            "en",
-            "fr",
-            "bar",
-            "nds",
-            "nl",
-            "it",
-            "es",
-            "ru",
-            "vmf",
-            "tr",
-            "gsw",
-            "da",
-            "hr",
-            "ku-latn",
-            "el",
-            "ksh",
-            "pl",
-            "hsb",
-            "frr",
-            "dsb",
-            "stq",
-            "pfl"
-        ],
-        "DG": [
-            "en"
-        ],
-        "DJ": [
-            "aa",
-            "so",
-            "ar",
-            "fr"
-        ],
-        "DK": [
-            "da",
-            "en",
-            "de",
-            "sv",
-            "fo",
-            "kl",
-            "jut"
-        ],
-        "DM": [
-            "en"
-        ],
-        "DO": [
-            "es",
-            "en"
-        ],
-        "DZ": [
-            "arq",
-            "ar",
-            "fr",
-            "kab",
-            "en"
-        ],
-        "EA": [
-            "es"
-        ],
-        "EC": [
-            "es",
-            "qu",
-            "qug"
-        ],
-        "EE": [
-            "et",
-            "ru",
-            "en",
-            "fi",
-            "vro"
-        ],
-        "EG": [
-            "ar",
-            "arz",
-            "en",
-            "el"
-        ],
-        "EH": [
-            "ar"
-        ],
-        "ER": [
-            "ti",
-            "en",
-            "tig",
-            "ar",
-            "aa",
-            "byn"
-        ],
-        "ES": [
-            "es",
-            "en",
-            "ca",
-            "gl",
-            "eu",
-            "ast",
-            "ext",
-            "an"
-        ],
-        "ET": [
-            "en",
-            "am",
-            "om",
-            "so",
-            "ti",
-            "aa"
-        ],
-        "FI": [
-            "fi",
-            "en",
-            "sv",
-            "de",
-            "ru",
-            "et",
-            "rmf",
-            "se",
-            "smn",
-            "sms"
-        ],
-        "FJ": [
-            "en",
-            "hi",
-            "hif",
-            "fj",
-            "rtm"
-        ],
-        "FK": [
-            "en"
-        ],
-        "FM": [
-            "en"
-        ],
-        "FO": [
-            "fo"
-        ],
-        "FR": [
-            "fr",
-            "en",
-            "es",
-            "de",
-            "oc",
-            "it",
-            "pt",
-            "pcd",
-            "gsw",
-            "br",
-            "co",
-            "ca",
-            "nl",
-            "eu",
-            "frp",
-            "ia"
-        ],
-        "GA": [
-            "fr"
-        ],
-        "GB": [
-            "en",
-            "fr",
-            "de",
-            "sco",
-            "pa-guru",
-            "cy",
-            "bn",
-            "zh-hant",
-            "zh",
-            "syl",
-            "el",
-            "it",
-            "ks-arab",
-            "gd",
-            "yi",
-            "ml",
-            "ga",
-            "kw"
-        ],
-        "GD": [
-            "en"
-        ],
-        "GE": [
-            "ka",
-            "xmf",
-            "ru",
-            "hy",
-            "ab",
-            "os",
-            "ku-latn"
-        ],
-        "GF": [
-            "fr",
-            "gcr",
-            "zh-hant",
-            "zh"
-        ],
-        "GG": [
-            "en"
-        ],
-        "GH": [
-            "ak",
-            "en",
-            "ee",
-            "gur",
-            "gaa",
-            "ha",
-            "ff"
-        ],
-        "GI": [
-            "en",
-            "es"
-        ],
-        "GL": [
-            "kl",
-            "da"
-        ],
-        "GM": [
-            "en",
-            "ff"
-        ],
-        "GN": [
-            "fr",
-            "ff",
-            "nqo"
-        ],
-        "GP": [
-            "fr"
-        ],
-        "GQ": [
-            "es",
-            "fan",
-            "fr",
-            "pt"
-        ],
-        "GR": [
-            "el",
-            "en",
-            "fr",
-            "de",
-            "pnt",
-            "mk",
-            "tr",
-            "bg",
-            "sq",
-            "tsd"
-        ],
-        "GT": [
-            "es",
-            "quc"
-        ],
-        "GU": [
-            "en",
-            "ch"
-        ],
-        "GW": [
-            "pt",
-            "ff"
-        ],
-        "GY": [
-            "en"
-        ],
-        "HK": [
-            "zh-hant",
-            "zh",
-            "yue",
-            "en"
-        ],
-        "HN": [
-            "es",
-            "en"
-        ],
-        "HR": [
-            "hr",
-            "en",
-            "it"
-        ],
-        "HT": [
-            "ht",
-            "fr"
-        ],
-        "HU": [
-            "hu",
-            "en",
-            "de",
-            "fr",
-            "ro",
-            "hr",
-            "sk",
-            "sl"
-        ],
-        "IC": [
-            "es"
-        ],
-        "ID": [
-            "id",
-            "jv",
-            "su",
-            "mad",
-            "ms",
-            "min",
-            "bew",
-            "ban",
-            "bug",
-            "bjn",
-            "ace",
-            "ms-arab",
-            "bbc-latn",
-            "zh-hant",
-            "zh",
-            "rej",
-            "gor",
-            "sly",
-            "mwv"
-        ],
-        "IE": [
-            "en",
-            "ga",
-            "fr"
-        ],
-        "IL": [
-            "he",
-            "en",
-            "ar",
-            "ru",
-            "ro",
-            "yi",
-            "pl",
-            "lad-latn",
-            "hu",
-            "am",
-            "ti",
-            "ml"
-        ],
-        "IM": [
-            "en",
-            "gv"
-        ],
-        "IN": [
-            "hi",
-            "en",
-            "bn",
-            "te",
-            "mr",
-            "ta",
-            "ur",
-            "gu",
-            "kn",
-            "ml",
-            "or",
-            "pa-guru",
-            "bho",
-            "awa",
-            "as",
-            "mai",
-            "hne",
-            "ne",
-            "sat",
-            "ks-arab",
-            "gom-deva",
-            "gbm",
-            "sd",
-            "doi",
-            "tcy",
-            "brx",
-            "mni",
-            "hoc",
-            "saz",
-            "ccp",
-            "bfq",
-            "njo",
-            "bo",
-            "bpy",
-            "bft",
-            "sa",
-            "dv",
-            "dz"
-        ],
-        "IO": [
-            "en"
-        ],
-        "IQ": [
-            "ar",
-            "en",
-            "ckb",
-            "az-arab",
-            "az-latn",
-            "fa",
-            "lrc"
-        ],
-        "IR": [
-            "fa",
-            "az-arab",
-            "az-latn",
-            "mzn",
-            "glk",
-            "ckb",
-            "sdh",
-            "tk",
-            "lrc",
-            "ar",
-            "bqi",
-            "luz",
-            "lki",
-            "bgn",
-            "hy",
-            "ps",
-            "ka",
-            "gbz",
-            "kk-arab",
-            "kk-cyrl"
-        ],
-        "IS": [
-            "is",
-            "da"
-        ],
-        "IT": [
-            "it",
-            "en",
-            "fr",
-            "sc",
-            "de",
-            "vec",
-            "nap",
-            "lij",
-            "scn",
-            "sl",
-            "sdc",
-            "fur",
-            "egl",
-            "ca",
-            "el",
-            "lmo",
-            "pms",
-            "hr",
-            "rgn"
-        ],
-        "JE": [
-            "en"
-        ],
-        "JM": [
-            "en",
-            "jam"
-        ],
-        "JO": [
-            "ar",
-            "en"
-        ],
-        "JP": [
-            "ja",
-            "ryu",
-            "ko"
-        ],
-        "KE": [
-            "sw",
-            "en",
-            "ki",
-            "so",
-            "pko",
-            "om",
-            "ar",
-            "pa-guru",
-            "gu"
-        ],
-        "KG": [
-            "ky",
-            "ru"
-        ],
-        "KH": [
-            "km"
-        ],
-        "KI": [
-            "en"
-        ],
-        "KM": [
-            "ar",
-            "fr"
-        ],
-        "KN": [
-            "en"
-        ],
-        "KP": [
-            "ko"
-        ],
-        "KR": [
-            "ko"
-        ],
-        "KW": [
-            "ar"
-        ],
-        "KY": [
-            "en"
-        ],
-        "KZ": [
-            "ru",
-            "kk-cyrl",
-            "en",
-            "de",
-            "ug-cyrl",
-            "ug-arab"
-        ],
-        "LA": [
-            "lo"
-        ],
-        "LB": [
-            "ar",
-            "en",
-            "hy",
-            "ku-arab",
-            "ku-latn",
-            "fr"
-        ],
-        "LC": [
-            "en"
-        ],
-        "LI": [
-            "de",
-            "gsw"
-        ],
-        "LK": [
-            "si",
-            "ta",
-            "en"
-        ],
-        "LR": [
-            "en",
-            "vai",
-            "ff"
-        ],
-        "LS": [
-            "st",
-            "en",
-            "zu",
-            "ss",
-            "xh"
-        ],
-        "LT": [
-            "lt",
-            "ru",
-            "en",
-            "de",
-            "sgs"
-        ],
-        "LU": [
-            "fr",
-            "lb",
-            "de",
-            "en",
-            "pt"
-        ],
-        "LV": [
-            "lv",
-            "en",
-            "ru",
-            "ltg"
-        ],
-        "LY": [
-            "ar"
-        ],
-        "MA": [
-            "ary",
-            "ar",
-            "zgh",
-            "fr",
-            "en",
-            "tzm",
-            "shi-latn",
-            "rif",
-            "es"
-        ],
-        "MC": [
-            "fr"
-        ],
-        "MD": [
-            "ro",
-            "uk",
-            "bg",
-            "gag",
-            "ru"
-        ],
-        "ME": [
-            "sr-latn",
-            "sr-cyrl",
-            "sq"
-        ],
-        "MF": [
-            "fr"
-        ],
-        "MG": [
-            "mg",
-            "fr",
-            "en"
-        ],
-        "MH": [
-            "en",
-            "mh"
-        ],
-        "MK": [
-            "mk",
-            "sq",
-            "tr"
-        ],
-        "ML": [
-            "bm",
-            "fr",
-            "ses",
-            "ar"
-        ],
-        "MM": [
-            "my",
-            "shn",
-            "kac",
-            "mnw"
-        ],
-        "MN": [
-            "mn",
-            "kk-arab",
-            "kk-cyrl",
-            "zh",
-            "ru",
-            "ug-cyrl",
-            "ug-arab"
-        ],
-        "MO": [
-            "zh-hant",
-            "zh",
-            "pt",
-            "en"
-        ],
-        "MP": [
-            "en",
-            "ch"
-        ],
-        "MQ": [
-            "fr"
-        ],
-        "MR": [
-            "ar",
-            "fr",
-            "ff",
-            "wo"
-        ],
-        "MS": [
-            "en"
-        ],
-        "MT": [
-            "mt",
-            "en",
-            "it",
-            "fr"
-        ],
-        "MU": [
-            "mfe",
-            "en",
-            "bho",
-            "ur",
-            "fr",
-            "ta"
-        ],
-        "MV": [
-            "dv"
-        ],
-        "MW": [
-            "en",
-            "ny",
-            "tum",
-            "zu"
-        ],
-        "MX": [
-            "es",
-            "en",
-            "yua",
-            "sei"
-        ],
-        "MY": [
-            "ms",
-            "en",
-            "zh",
-            "ta",
-            "jv",
-            "dtp",
-            "ml",
-            "bug",
-            "bjn"
-        ],
-        "MZ": [
-            "pt",
-            "ts",
-            "ny",
-            "sw",
-            "zu"
-        ],
-        "NA": [
-            "af",
-            "kj",
-            "ng",
-            "hz",
-            "en",
-            "de",
-            "tn"
-        ],
-        "NC": [
-            "fr"
-        ],
-        "NE": [
-            "ha",
-            "fr",
-            "ar",
-            "ff"
-        ],
-        "NF": [
-            "en"
-        ],
-        "NG": [
-            "en",
-            "ha",
-            "ig",
-            "yo",
-            "ha-arab",
-            "kcg",
-            "ar",
-            "ff"
-        ],
-        "NI": [
-            "es"
-        ],
-        "NL": [
-            "nl",
-            "en",
-            "de",
-            "fr",
-            "nds",
-            "li",
-            "fy",
-            "id",
-            "zea",
-            "rif",
-            "tr"
-        ],
-        "NO": [
-            "nb",
-            "nn",
-            "se"
-        ],
-        "NP": [
-            "ne",
-            "mai",
-            "bho",
-            "new",
-            "en",
-            "dty",
-            "awa",
-            "hi",
-            "bo",
-            "bn"
-        ],
-        "NR": [
-            "en",
-            "na"
-        ],
-        "NU": [
-            "en",
-            "niu"
-        ],
-        "NZ": [
-            "en",
-            "mi"
-        ],
-        "OM": [
-            "ar",
-            "fa"
-        ],
-        "PA": [
-            "es",
-            "en",
-            "zh-hant",
-            "zh"
-        ],
-        "PE": [
-            "es",
-            "qu",
-            "ay"
-        ],
-        "PF": [
-            "fr",
-            "ty",
-            "zh-hant",
-            "zh"
-        ],
-        "PG": [
-            "tpi",
-            "en",
-            "ho"
-        ],
-        "PH": [
-            "en",
-            "tl",
-            "es",
-            "ceb",
-            "ilo",
-            "hil",
-            "war",
-            "pam",
-            "pag",
-            "zh-hant",
-            "zh",
-            "cps",
-            "krj",
-            "bto"
-        ],
-        "PK": [
-            "ur",
-            "pa-guru",
-            "en",
-            "ps",
-            "sd",
-            "skr-arab",
-            "brh",
-            "fa",
-            "bgn",
-            "tg-cyrl",
-            "bft",
-            "khw",
-            "ks-arab"
-        ],
-        "PL": [
-            "pl",
-            "en",
-            "de",
-            "ru",
-            "szl",
-            "be",
-            "uk",
-            "csb",
-            "sli",
-            "lt"
-        ],
-        "PM": [
-            "fr",
-            "en"
-        ],
-        "PN": [
-            "en"
-        ],
-        "PR": [
-            "es",
-            "en"
-        ],
-        "PS": [
-            "ar"
-        ],
-        "PT": [
-            "pt",
-            "en",
-            "fr",
-            "es",
-            "gl"
-        ],
-        "PW": [
-            "en"
-        ],
-        "PY": [
-            "gn",
-            "es",
-            "de"
-        ],
-        "QA": [
-            "ar",
-            "fa",
-            "ml"
-        ],
-        "RE": [
-            "fr",
-            "rcf",
-            "ta"
-        ],
-        "RO": [
-            "ro",
-            "en",
-            "fr",
-            "es",
-            "hu",
-            "de",
-            "tr",
-            "sr-latn",
-            "sr-cyrl",
-            "bg",
-            "el",
-            "pl"
-        ],
-        "RS": [
-            "sr-cyrl",
-            "sr-latn",
-            "sq",
-            "hu",
-            "ro",
-            "hr",
-            "sk",
-            "uk"
-        ],
-        "RU": [
-            "ru",
-            "tt",
-            "ba",
-            "cv",
-            "hy",
-            "ce",
-            "av",
-            "udm",
-            "mhr",
-            "sah",
-            "os",
-            "kbd",
-            "myv",
-            "dar",
-            "mdf",
-            "kum",
-            "kv",
-            "lez",
-            "krc",
-            "inh",
-            "tyv",
-            "az-cyrl",
-            "az-latn",
-            "ady",
-            "krl",
-            "lbe",
-            "koi",
-            "mrj",
-            "alt",
-            "fi",
-            "sr-latn",
-            "sr-cyrl",
-            "vep",
-            "mn",
-            "izh",
-            "vot",
-            "cu"
-        ],
-        "RW": [
-            "rw",
-            "en",
-            "fr"
-        ],
-        "SA": [
-            "ar"
-        ],
-        "SB": [
-            "en"
-        ],
-        "SC": [
-            "fr",
-            "en"
-        ],
-        "SD": [
-            "ar",
-            "en",
-            "ha-arab",
-            "ha"
-        ],
-        "SE": [
-            "sv",
-            "en",
-            "fi",
-            "fit",
-            "se",
-            "yi",
-            "smj",
-            "sma",
-            "ia"
-        ],
-        "SG": [
-            "en",
-            "zh",
-            "ms",
-            "ta",
-            "ml",
-            "pa-guru"
-        ],
-        "SH": [
-            "en"
-        ],
-        "SI": [
-            "sl",
-            "hr",
-            "en",
-            "de",
-            "hu",
-            "it"
-        ],
-        "SJ": [
-            "nb",
-            "ru"
-        ],
-        "SK": [
-            "sk",
-            "cs",
-            "en",
-            "de",
-            "hu",
-            "uk",
-            "pl"
-        ],
-        "SL": [
-            "kri",
-            "en",
-            "ff"
-        ],
-        "SM": [
-            "it",
-            "eo"
-        ],
-        "SN": [
-            "wo",
-            "fr",
-            "ff"
-        ],
-        "SO": [
-            "so",
-            "ar",
-            "sw",
-            "om"
-        ],
-        "SR": [
-            "nl",
-            "srn",
-            "zh-hant",
-            "zh"
-        ],
-        "SS": [
-            "ar",
-            "en",
-            "nus"
-        ],
-        "ST": [
-            "pt"
-        ],
-        "SV": [
-            "es"
-        ],
-        "SX": [
-            "en",
-            "es",
-            "nl"
-        ],
-        "SY": [
-            "ar",
-            "ku-latn",
-            "fr",
-            "hy"
-        ],
-        "SZ": [
-            "en",
-            "ss",
-            "zu",
-            "ts"
-        ],
-        "TA": [
-            "en"
-        ],
-        "TC": [
-            "en"
-        ],
-        "TD": [
-            "fr",
-            "ar"
-        ],
-        "TF": [
-            "fr"
-        ],
-        "TG": [
-            "fr",
-            "ee"
-        ],
-        "TH": [
-            "th",
-            "en",
-            "nod",
-            "zh-hant",
-            "zh",
-            "mnw",
-            "shn"
-        ],
-        "TJ": [
-            "tg-cyrl",
-            "ru",
-            "fa",
-            "ar"
-        ],
-        "TK": [
-            "en"
-        ],
-        "TL": [
-            "pt",
-            "tet"
-        ],
-        "TM": [
-            "tk",
-            "ru",
-            "uz",
-            "ku-latn"
-        ],
-        "TN": [
-            "aeb-arab",
-            "ar",
-            "fr"
-        ],
-        "TO": [
-            "to",
-            "en"
-        ],
-        "TR": [
-            "tr",
-            "en",
-            "ku-latn",
-            "kbd",
-            "az-latn",
-            "az-arab",
-            "ar",
-            "bg",
-            "ady",
-            "kiu",
-            "hy",
-            "ka",
-            "sr-latn",
-            "sr-cyrl",
-            "lzz",
-            "sq",
-            "ab",
-            "el",
-            "tru",
-            "uz",
-            "ky",
-            "kk-cyrl"
-        ],
-        "TT": [
-            "en",
-            "es"
-        ],
-        "TV": [
-            "en"
-        ],
-        "TW": [
-            "zh-hant",
-            "zh",
-            "trv"
-        ],
-        "TZ": [
-            "sw",
-            "en",
-            "lag"
-        ],
-        "UA": [
-            "uk",
-            "ru",
-            "pl",
-            "yi",
-            "rue",
-            "be",
-            "crh",
-            "ro",
-            "bg",
-            "tr",
-            "hu",
-            "el"
-        ],
-        "UG": [
-            "sw",
-            "lg",
-            "en",
-            "rw",
-            "hi"
-        ],
-        "UM": [
-            "en"
-        ],
-        "US": [
-            "en",
-            "es",
-            "zh-hant",
-            "zh",
-            "fr",
-            "de",
-            "tl",
-            "it",
-            "vi",
-            "ko",
-            "ru",
-            "nv",
-            "yi",
-            "pdc",
-            "haw",
-            "frc",
-            "chr",
-            "esu",
-            "cho",
-            "lkt",
-            "ik",
-            "mus"
-        ],
-        "UY": [
-            "es"
-        ],
-        "UZ": [
-            "uz",
-            "ru",
-            "kaa",
-            "tr"
-        ],
-        "VA": [
-            "it",
-            "la"
-        ],
-        "VC": [
-            "en"
-        ],
-        "VE": [
-            "es"
-        ],
-        "VG": [
-            "en"
-        ],
-        "VI": [
-            "en"
-        ],
-        "VN": [
-            "vi",
-            "zh-hant",
-            "zh"
-        ],
-        "VU": [
-            "bi",
-            "en",
-            "fr"
-        ],
-        "WF": [
-            "wls",
-            "fr"
-        ],
-        "WS": [
-            "sm",
-            "en"
-        ],
-        "XK": [
-            "sq",
-            "aln",
-            "sr-cyrl",
-            "sr-latn"
-        ],
-        "YE": [
-            "ar",
-            "en"
-        ],
-        "YT": [
-            "swb",
-            "fr",
-            "sw"
-        ],
-        "ZA": [
-            "en",
-            "zu",
-            "xh",
-            "af",
-            "nso",
-            "tn",
-            "st",
-            "ts",
-            "ss",
-            "ve",
-            "hi",
-            "nr",
-            "sw"
-        ],
-        "ZM": [
-            "en",
-            "ny",
-            "loz"
-        ],
-        "ZW": [
-            "sn",
-            "en",
-            "nd",
-            "ny",
-            "ve",
-            "tn"
-        ]
-    }
-}
-
-		}( jQuery ) );
-	}, {} ]
-}, {}, [ 1 ] );
