+++ test/index.html
@@ -5,18 +5,14 @@
 <title>jQuery ULS Test Suite</title>
 
 <!-- External libs -->
-<script src="https://code.jquery.com/jquery-1.11.1.min.js"></script>
+<script src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
 
 <!-- Source of our libs -->
-<script src="../src/jquery.uls.data.js"></script>
-<script src="../src/jquery.uls.data.utils.js"></script>
-<script src="../src/jquery.uls.lcd.js"></script>
-<script src="../src/jquery.uls.languagefilter.js"></script>
-<script src="../src/jquery.uls.core.js"></script>
+<script src="../dist/jquery.uls.js"></script>
 
 <!-- Test framework -->
-<link rel="stylesheet" href="https://code.jquery.com/qunit/qunit-1.14.0.css">
-<script src="https://code.jquery.com/qunit/qunit-1.14.0.js"></script>
+<link rel="stylesheet" href="https://code.jquery.com/qunit/qunit-2.15.0.css">
+<script src="https://code.jquery.com/qunit/qunit-2.15.0.js"></script>
 
 <!-- Test Suite -->
 <script src="jquery.uls.test.js"></script>
